Add common "init" code
This initializes the UART and turns on the ACT LED.

The idea is that this will allow other code to use the UART for logging and the ACT LED as an indicator, and to assume they've been initialized.
The UART interface is particularly invaluable for debugging, and I intend to soon instrument almost everything with logging.