Add "general notes"
These are notes not relating to any specific implementation, or relating to implementations that don't yet exist.

An example in the former category is the pin layout on the USB-to-UART debug cable.
An example in the latter category is the AHT20 datasheet and misc info, which will be moved to `src/devices/aht20.c` once I get around to implementing that.