Switch to EL2 in boot.s
This allows us to set a vector table for EL2 and EL3 as well and maybe catch some errors taken to those ELs that we wouldn't otherwise catch.
Since we read `e{sr,lr}_el1` in the exception handler those values probably won't be correct, but at least we'll get some indication of the error rather than a triple fault (or whatever it's called in ARM).