Implement "first free block" optimization
When we allocate a block, we know that it won't be free, so there's no reason to make subsequent malloc calls traverse it.
Thus, if the block we allocated was the current head block, we can set the head block to the block after this one.

Conversely, when freeing a block, if it is before the current head, we can update the head.